pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "modular-agents"
version = "0.0.1"
description = "Modular agents"
package-mode = false
readme = "README.md"
license = "Apache-2.0"
authors = ["Martin Krasser <krasserm@googlemail.com>"]

[tool.poetry.dependencies]
accelerate = "^0.30.1"
bitsandbytes = "^0.43.1"
chromadb = "^0.5"
datasets = "^2.19"
jsonargparse= "^4.28"
jsonref = "^1.1"
langchain = "^0.1.20"
langchain-experimental = "^0.0.58"
notebook = "^7.1"
numexpr = "^2.8"
openai = "^1.30"
pandas = "^1.3"
peft = "^0.11.1"
python = "^3.10,<3.13"
python-dotenv = "^1.0"
requests = "^2.32"
safetensors = "^0.4.3"
sentence-transformers = "^2.7"
torch = "~2.2"
transformers = "^4.41"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7"
invoke = "^2.2"
pytest = "^8.2"

[tool.ruff]
line-length = 120
exclude = ["gba/client/grammar.py"]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]
